avoid delete issue

+++ app/Http/Controllers/CustomerController.php
@@ -77,6 +77,9 @@ class CustomerController extends Controller
 
     public function delete($id){
         $item = Customer::find($id);
+        if(!$item){
+            return back()->withErrors(["delete" => __('page.something_went_wrong')]);
+        }
         $item->delete();
         return back()->with("success", __('page.deleted_successfully'));
     }
