+++ app/Http/Controllers/StoreController.php
@@ -49,6 +49,9 @@ class StoreController extends Controller
 
     public function delete($id){
         $item = Store::find($id);
+        if(!$item){
+            return back()->withErrors(["delete" => __('page.something_went_wrong')]);
+        }
         $item->delete();
         return back()->with("success", __('page.deleted_successfully'));
     }
